% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ab_kin_normal}
\alias{ab_kin_normal}
\title{Base boosting fast}
\usage{
ab_kin_normal(
  predicted_titres,
  theta,
  infection_info,
  vaccination_info,
  setup_data,
  indexing,
  antigenic_maps,
  other_pars
)
}
\description{
A fast implementation of the basic boosting function, giving predicted titres for a number of samples for one individual. Note that this version attempts to minimise memory allocations.
}
\seealso{
\code{\link{titre_data_fast}}

Other boosting_functions: 
\code{\link{ab_kin_vac}()}
}
\concept{boosting_functions}
