% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_funcs.R
\name{make_antigenic_maps_default_n}
\alias{make_antigenic_maps_default_n}
\title{Function to make the antigenic maps (no vaccination invovled)}
\usage{
make_antigenic_maps_default_n(antigenic_map_melted, theta)
}
\description{
Function to make the antigenic maps (no vaccination invovled)
}
